Thermal_Diffusion_Project/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
accelerate>=0.20.0
pyyaml
tqdm
matplotlib
pillow
